Extra table pagination bar in a component
